+++ lib/Payever/Payments/PaymentsApiClient.php
@@ -26,6 +26,7 @@ use Payever\Sdk\Payments\Http\RequestEntity\CreatePaymentV2Request;
 use Payever\Sdk\Payments\Http\RequestEntity\CreatePaymentV3Request;
 use Payever\Sdk\Payments\Http\RequestEntity\ListPaymentsRequest;
 use Payever\Sdk\Payments\Http\RequestEntity\RefundPaymentRequest;
+use Payever\Sdk\Payments\Http\RequestEntity\SettlePaymentRequest;
 use Payever\Sdk\Payments\Http\RequestEntity\ShippingGoodsPaymentRequest;
 use Payever\Sdk\Payments\Http\RequestEntity\RefundItemsPaymentRequest;
 use Payever\Sdk\Payments\Http\RequestEntity\CancelItemsPaymentRequest;
@@ -44,10 +45,12 @@ use Payever\Sdk\Payments\Http\ResponseEntity\LatePaymentsResponse;
 use Payever\Sdk\Payments\Http\ResponseEntity\ListPaymentOptionsResponse;
 use Payever\Sdk\Payments\Http\ResponseEntity\ListPaymentOptionsWithVariantsResponse;
 use Payever\Sdk\Payments\Http\ResponseEntity\ListPaymentsResponse;
+use Payever\Sdk\Payments\Http\ResponseEntity\PaymentSettingsResponse;
 use Payever\Sdk\Payments\Http\ResponseEntity\RefundPaymentResponse;
 use Payever\Sdk\Payments\Http\ResponseEntity\RemindPaymentResponse;
 use Payever\Sdk\Payments\Http\ResponseEntity\RetrieveApiCallResponse;
 use Payever\Sdk\Payments\Http\ResponseEntity\RetrievePaymentResponse;
+use Payever\Sdk\Payments\Http\ResponseEntity\SettlePaymentResponse;
 use Payever\Sdk\Payments\Http\ResponseEntity\SubmitPaymentResponse;
 use Payever\Sdk\Payments\Http\ResponseEntity\ShippingGoodsPaymentResponse;
 use Payever\Sdk\Payments\Http\ResponseEntity\CompanySearchResponse;
@@ -77,13 +80,19 @@ class PaymentsApiClient extends CommonApiClient implements PaymentsApiClientInte
     const SUB_URL_CANCEL_PAYMENT = 'api/payment/cancel/%s';
     const SUB_URL_CLAIM_PAYMENT = 'api/payment/claim/%s';
     const SUB_URL_INVOICE_PAYMENT = 'api/payment/invoice/%s';
+    const SUB_URL_SETTLE_PAYMENT = 'api/payment/settle/%s';
     const SUB_URL_RETRIEVE_API_CALL = 'api/%s';
     const SUB_URL_LIST_PAYMENT_OPTIONS = 'api/shop/oauth/%s/payment-options/%s';
     const SUB_URL_LIST_PAYMENT_OPTIONS_VARIANTS = 'api/shop/oauth/%s/payment-options/variants/%s';
+    const SUB_URL_SETTINGS = 'api/shop/%s/settings/%s';
     const SUB_URL_TRANSACTION = 'api/rest/v1/transactions/%s';
 
     const SUB_URL_COMPANY_SEARCH = 'api/b2b/search';
     const SUB_URL_COMPANY_SEARCH_CREDIT = 'api/b2b/search/credit';
+
+    const ACCOUNT_B2B_TYPE = 'b2b';
+    const ACCOUNT_B2C_TYPE = 'b2c';
+    const ACCOUNT_MIXED_TYPE = 'mixed';
 
     /**
      * {@inheritdoc}
@@ -550,6 +559,27 @@ class PaymentsApiClient extends CommonApiClient implements PaymentsApiClientInte
      *
      * @throws \Exception
      */
+    public function settlePaymentRequest($paymentId, SettlePaymentRequest $paymentRequest = null)
+    {
+        $this->configuration->assertLoaded();
+
+        $request = RequestBuilder::post($this->getSettlePaymentURL($paymentId))
+            ->addRawHeader(
+                $this->getToken(OauthToken::SCOPE_PAYMENT_ACTIONS)->getAuthorizationString()
+            )
+            ->contentTypeIsJson()
+            ->setRequestEntity($paymentRequest ?: new SettlePaymentRequest())
+            ->setResponseEntity(new SettlePaymentResponse())
+            ->build();
+
+        return $this->executeRequest($request);
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @throws \Exception
+     */
     public function retrieveApiCallRequest($callId)
     {
         $this->configuration->assertLoaded();
@@ -602,6 +632,47 @@ class PaymentsApiClient extends CommonApiClient implements PaymentsApiClientInte
             ->build();
 
         return $this->executeRequest($request, OauthToken::SCOPE_PAYMENT_INFO);
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @throws \Exception
+     */
+    public function paymentSettingsRequest($businessUuid = '', $channel = '')
+    {
+        $businessUuid = $businessUuid ?: $this->getConfiguration()->getBusinessUuid();
+        $channel = $channel ?: $this->getConfiguration()->getChannelSet();
+
+        $request = RequestBuilder::get($this->getPaymentSettingsURL($businessUuid, $channel))
+                                 ->setResponseEntity(new PaymentSettingsResponse())
+                                 ->build();
+
+        return $this->executeRequest($request);
+    }
+
+    /**
+     * Returns if B2b Search is Active
+     *
+     * @param string $businessUuid
+     * @param string $channel
+     *
+     * @return bool
+     * @throws \Exception
+     */
+    public function isB2bSearchActive($businessUuid = '', $channel = '')
+    {
+        /** @var PaymentSettingsResponse $response */
+        $response = $this->paymentSettingsRequest($businessUuid, $channel)->getResponseEntity();
+
+        if (
+            ($response->getType() === self::ACCOUNT_B2B_TYPE || $response->getType() === self::ACCOUNT_MIXED_TYPE)
+            && $response->getB2bSearch()
+        ) {
+            return true;
+        }
+
+        return false;
     }
 
     /**
@@ -824,6 +895,18 @@ class PaymentsApiClient extends CommonApiClient implements PaymentsApiClientInte
     }
 
     /**
+     * Returns URL for Settle Payment path
+     *
+     * @param string $paymentId
+     *
+     * @return string
+     */
+    protected function getSettlePaymentURL($paymentId)
+    {
+        return $this->getBaseUrl() . sprintf(self::SUB_URL_SETTLE_PAYMENT, $paymentId);
+    }
+
+    /**
      * Returns URL for Retrieve API Call path
      *
      * @param string $callId
@@ -865,6 +948,19 @@ class PaymentsApiClient extends CommonApiClient implements PaymentsApiClientInte
         return $this->getBaseUrl()
             . sprintf(self::SUB_URL_LIST_PAYMENT_OPTIONS_VARIANTS, $businessUuid, $channel)
             . (empty($params) ? '' : '?' . http_build_query($params));
+    }
+
+    /**
+     * Returns payment settings
+     *
+     * @param string $businessUuid
+     * @param string $channel
+     *
+     * @return string
+     */
+    protected function getPaymentSettingsURL($businessUuid, $channel)
+    {
+        return $this->getBaseUrl() . sprintf(self::SUB_URL_SETTINGS, $businessUuid, $channel);
     }
 
     /**